// src/gen/types/Device.ts
export type Device = {
  /**
   * @type string
   */
  id: string
  /**
   * @type string
   */
  uid: string
  /**
   * @type number, double
   */
  longitude: number
  /**
   * @type number, double
   */
  latitude: number
  /**
   * @type string
   */
  deviceId: string
  /**
   * @type strin
   */
  deviceToken
}